refactor(hooks): split config and executor modules
Move data-only hook configuration into hooks/config.rs and process execution, timeout, and decision parsing into hooks/executor.rs while preserving crate::hooks re-exports.

Closes #4082.

Harvested from PR #4087 by @cyq1017
