+++ crates/tui/src/hooks/config.rs
@@ -1,0 +1,244 @@
+use serde::{Deserialize, Serialize};
+use std::path::{Path, PathBuf};
+
+/// Events that can trigger hook execution
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum HookEvent {
+    /// Triggered when a new session starts
+    SessionStart,
+    /// Triggered when a session ends (quit, Ctrl+C)
+    SessionEnd,
+    /// Triggered before a user message is sent to the LLM
+    MessageSubmit,
+    /// Triggered before a tool is executed
+    ToolCallBefore,
+    /// Triggered after a tool completes (success or failure)
+    ToolCallAfter,
+    /// Triggered when the user changes modes (Plan, Agent, Yolo)
+    ModeChange,
+    /// Triggered when an error occurs
+    OnError,
+    /// Triggered after a turn completes and post-turn state has been updated
+    TurnEnd,
+    /// Triggered when a sub-agent is spawned
+    SubagentSpawn,
+    /// Triggered when a sub-agent reaches a terminal state
+    SubagentComplete,
+    /// Triggered immediately before each `exec_shell` invocation. The hook's
+    /// stdout is parsed as `KEY=VALUE\n` lines and merged on top of the
+    /// shell command's environment — useful for ephemeral credentials,
+    /// per-skill PATH adjustments, or short-lived tokens (#456). Hooks that
+    /// fail or time out are logged but do *not* abort the shell call; they
+    /// simply contribute no env vars.
+    ShellEnv,
+}
+
+impl HookEvent {
+    /// Get string representation for environment variable
+    #[allow(dead_code)] // Used in tests and future hook dispatch
+    pub fn as_str(self) -> &'static str {
+        match self {
+            HookEvent::SessionStart => "session_start",
+            HookEvent::SessionEnd => "session_end",
+            HookEvent::MessageSubmit => "message_submit",
+            HookEvent::ToolCallBefore => "tool_call_before",
+            HookEvent::ToolCallAfter => "tool_call_after",
+            HookEvent::ModeChange => "mode_change",
+            HookEvent::OnError => "on_error",
+            HookEvent::TurnEnd => "turn_end",
+            HookEvent::SubagentSpawn => "subagent_spawn",
+            HookEvent::SubagentComplete => "subagent_complete",
+            HookEvent::ShellEnv => "shell_env",
+        }
+    }
+}
+
+/// Condition for when a hook should run
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type", rename_all = "snake_case")]
+#[derive(Default)]
+pub enum HookCondition {
+    /// Always run this hook
+    #[default]
+    Always,
+    /// Only run for specific tool names
+    ToolName {
+        /// Tool name to match (e.g., "`exec_shell`", "`write_file`")
+        name: String,
+    },
+    /// Only run for specific tool categories
+    ToolCategory {
+        /// Category: "safe", "`file_write`", "shell"
+        category: String,
+    },
+    /// Only run in specific modes
+    Mode {
+        /// Mode: "plan", "agent", "yolo"
+        mode: String,
+    },
+    /// Only run when exit code matches (for `ToolCallAfter`)
+    ExitCode {
+        /// Exit code to match
+        code: i32,
+    },
+    /// Combine multiple conditions with AND
+    All { conditions: Vec<HookCondition> },
+    /// Combine multiple conditions with OR
+    Any { conditions: Vec<HookCondition> },
+}
+
+/// A single hook definition
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Hook {
+    /// The event that triggers this hook
+    pub event: HookEvent,
+
+    /// Shell command to execute (platform shell: `sh -c` on Unix, `cmd /C` on Windows)
+    pub command: String,
+
+    /// Optional condition for when this hook should run
+    #[serde(default)]
+    pub condition: Option<HookCondition>,
+
+    /// Timeout in seconds (default: 30)
+    #[serde(default = "default_timeout")]
+    pub timeout_secs: u64,
+
+    /// Run in background (don't wait for completion)
+    #[serde(default)]
+    pub background: bool,
+
+    /// Continue if this hook fails (default: true)
+    #[serde(default = "default_continue_on_error")]
+    pub continue_on_error: bool,
+
+    /// Optional name for logging/debugging
+    #[serde(default)]
+    pub name: Option<String>,
+}
+
+fn default_timeout() -> u64 {
+    30
+}
+
+fn default_continue_on_error() -> bool {
+    true
+}
+
+impl Hook {
+    /// Create a new hook with minimal configuration
+    #[allow(dead_code)] // Public builder API, used in tests
+    pub fn new(event: HookEvent, command: &str) -> Self {
+        Self {
+            event,
+            command: command.to_string(),
+            condition: None,
+            timeout_secs: 30,
+            background: false,
+            continue_on_error: true,
+            name: None,
+        }
+    }
+
+    /// Builder: set condition
+    #[allow(dead_code)] // Public builder API, used in tests
+    pub fn with_condition(mut self, condition: HookCondition) -> Self {
+        self.condition = Some(condition);
+        self
+    }
+
+    /// Builder: set timeout
+    #[allow(dead_code)] // Public builder API, used in tests
+    pub fn with_timeout(mut self, secs: u64) -> Self {
+        self.timeout_secs = secs;
+        self
+    }
+
+    /// Builder: run in background
+    #[allow(dead_code)] // Public builder API, used in tests
+    pub fn background(mut self) -> Self {
+        self.background = true;
+        self
+    }
+
+    /// Builder: set name
+    #[allow(dead_code)] // Public builder API, used in tests
+    pub fn with_name(mut self, name: &str) -> Self {
+        self.name = Some(name.to_string());
+        self
+    }
+}
+
+/// Configuration for hooks (loaded from config.toml)
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct HooksConfig {
+    /// List of hooks to execute
+    #[serde(default)]
+    pub hooks: Vec<Hook>,
+
+    /// Global enable/disable for all hooks
+    #[serde(default = "default_enabled")]
+    pub enabled: bool,
+
+    /// Global timeout override (applies if hook doesn't specify one)
+    #[serde(default)]
+    pub default_timeout_secs: Option<u64>,
+
+    /// Working directory for hook execution (default: workspace)
+    #[serde(default)]
+    pub working_dir: Option<PathBuf>,
+}
+
+fn default_enabled() -> bool {
+    true
+}
+
+impl HooksConfig {
+    /// Load global hooks merged with project-local `.codewhale/hooks.toml` (#3026).
+    ///
+    /// Project hooks are executable repository configuration, so they are only
+    /// honored after the workspace has been trusted in user-owned config.
+    /// Trusted project hooks are appended after global hooks.  A malformed
+    /// trusted project file logs a warning and falls back to global-only.
+    pub fn load_with_project(global: HooksConfig, workspace: &Path) -> HooksConfig {
+        let project_path = workspace.join(".codewhale").join("hooks.toml");
+        if !project_path.exists() || !workspace_allows_project_hooks(workspace) {
+            return global;
+        }
+        let Ok(contents) = std::fs::read_to_string(&project_path) else {
+            return global;
+        };
+        let project: HooksConfig = match toml::from_str(&contents) {
+            Ok(cfg) => cfg,
+            Err(e) => {
+                tracing::warn!(
+                    "Failed to parse project hooks at {}: {e}; falling back to global hooks only",
+                    project_path.display()
+                );
+                return global;
+            }
+        };
+        let mut merged = global;
+        merged.hooks.extend(project.hooks);
+        merged
+    }
+
+    /// Get hooks for a specific event
+    pub fn hooks_for_event(&self, event: HookEvent) -> Vec<&Hook> {
+        if !self.enabled {
+            return Vec::new();
+        }
+        self.hooks.iter().filter(|h| h.event == event).collect()
+    }
+
+    /// Check if hooks are configured and enabled
+    #[allow(dead_code)] // Public API for hook system consumers
+    pub fn has_hooks(&self) -> bool {
+        self.enabled && !self.hooks.is_empty()
+    }
+}
+
+fn workspace_allows_project_hooks(workspace: &Path) -> bool {
+    crate::config::is_workspace_trusted(workspace)
+}
